+++ include/PetriEngine/ExplicitColored/ColoredPetriNetState.h
@@ -3,7 +3,6 @@
 
 #include <queue>
 #include <utility>
-#include "PetriEngine/ExplicitColored/IntegerPackCodec.h"
 #include "ColoredPetriNetMarking.h"
 
 namespace PetriEngine::ExplicitColored {
@@ -125,70 +124,6 @@ namespace PetriEngine::ExplicitColored {
         uint32_t _currentIndex = 0;
         uint32_t _completedTransitions = 0;
     };
-
-    struct PossibleValues {
-            explicit PossibleValues(std::vector<Color_t> colors)
-                : colors(std::move(colors)), allColors(false) {}
-
-            explicit PossibleValues(const std::set<Color_t>& colors)
-                : colors(colors.begin(), colors.end()), allColors(false) {}
-
-            static PossibleValues getAll() {
-                PossibleValues rv(std::vector<Color_t> {});
-                rv.allColors = true;
-                return rv;
-            }
-
-            static PossibleValues getEmpty() {
-                PossibleValues rv(std::vector<Color_t> {});
-                rv.allColors = false;
-                return rv;
-            }
-
-            void sort() {
-                std::sort(colors.begin(), colors.end());
-            }
-
-            void intersect(const PossibleValues& other) {
-                if (other.allColors) {
-                    return;
-                }
-                if (allColors) {
-                    colors = other.colors;
-                    return;
-                }
-                std::vector<Color_t> newColors;
-                std::set_intersection(
-                    colors.cbegin(),
-                    colors.cend(),
-                    other.colors.cbegin(),
-                    other.colors.cend(),
-                    std::back_inserter(newColors)
-                );
-                colors = std::move(newColors);
-            }
-
-            void intersect(const std::set<Color_t>& other) {
-                if (allColors) {
-                    colors.clear();
-                    colors.insert(colors.begin(), other.cbegin(), other.cend());
-                    return;
-                }
-
-                std::vector<Color_t> newColors;
-                std::set_intersection(
-                    colors.cbegin(),
-                    colors.cend(),
-                    other.cbegin(),
-                    other.cend(),
-                    std::back_inserter(newColors)
-                );
-                colors = std::move(newColors);
-            }
-
-            std::vector<Color_t> colors;
-            bool allColors;
-        };
 }
 
 #endif //COLOREDPETRINETSTATE_H
