+++ include/PetriEngine/ExplicitColored/SuccessorGenerator/IntegerPackCodec.h
