+++ include/utils/MathExt.h
@@ -4,7 +4,7 @@
 /// maps value to [0,max) using modulo but when value is negative
 /// it is handled differently s.t. signed_wrap(-1, 5) = 4
 template <typename S, typename U>
-std::make_unsigned_t<U> signed_wrap(S value, U max) {
+std::make_unsigned_t<U> signedWrap(S value, U max) {
     static_assert(std::numeric_limits<S>::min() >= std::numeric_limits<int64_t>::min()
         && std::numeric_limits<S>::max() <= std::numeric_limits<int64_t>::max(), "S needs to be containable in int64_t");
     static_assert(std::numeric_limits<U>::min() >= std::numeric_limits<int64_t>::min()
@@ -14,12 +14,12 @@ std::make_unsigned_t<U> signed_wrap(S value, U max) {
 }
 
 template<typename U, typename O>
-std::make_unsigned_t<U> add_color_offset(U baseValue, O offset, U max) {
+std::make_unsigned_t<U> addColorOffset(U baseValue, O offset, U max) {
     static_assert(std::numeric_limits<U>::min() >= std::numeric_limits<int64_t>::min()
         && std::numeric_limits<U>::max() <= std::numeric_limits<int64_t>::max(), "U needs to be containable in int64_t");
     static_assert(std::numeric_limits<O>::min() >= std::numeric_limits<int64_t>::min()
         && std::numeric_limits<O>::max() <= std::numeric_limits<int64_t>::max(), "O needs to be containable in int64_t");
     static_assert(std::is_signed_v<O>, "O needs to be signed");
-    return signed_wrap(static_cast<int64_t>(baseValue) + static_cast<int64_t>(offset), max);
+    return signedWrap(static_cast<int64_t>(baseValue) + static_cast<int64_t>(offset), max);
 }
 #endif //MATHEXT_H
