+++ src/PetriEngine/ExplicitColored/ArcCompiler.cpp
@@ -1,6 +1,6 @@
-#include "PetriEngine/ExplicitColored/ArcCompiler.h"
+#include "PetriEngine/ExplicitColored/ExpressionCompilers/ArcCompiler.h"
 #include <PetriEngine/ExplicitColored/ExplicitErrors.h>
-#include "PetriEngine/ExplicitColored/VariableExtractorVisitor.h"
+#include "PetriEngine/ExplicitColored/Visitors/VariableExtractorVisitor.h"
 #include "utils/MathExt.h"
 
 namespace PetriEngine::ExplicitColored {
@@ -369,11 +369,11 @@ namespace PetriEngine::ExplicitColored {
                 const auto& parameterizedColor = sequence[i];
                 if (sequence[i].isVariable) {
                     colorSequence.push_back(
-                        add_color_offset(binding.getValue(parameterizedColor.value.variable), parameterizedColor.offset, _colorSizes[i])
+                        addColorOffset(binding.getValue(parameterizedColor.value.variable), parameterizedColor.offset, _colorSizes[i])
                     );
                 } else {
                     colorSequence.push_back(
-                        add_color_offset(parameterizedColor.value.color, parameterizedColor.offset, _colorSizes[i])
+                        addColorOffset(parameterizedColor.value.color, parameterizedColor.offset, _colorSizes[i])
                     );
                 }
             }
@@ -644,7 +644,7 @@ namespace PetriEngine::ExplicitColored {
                             hasVariable = true;
                             break;
                         }
-                        newColorSequence.push_back(add_color_offset( color.value.color, color.offset, expr->getColorMaxes()[i]));
+                        newColorSequence.push_back(addColorOffset( color.value.color, color.offset, expr->getColorMaxes()[i]));
                     }
                     if (!hasVariable) {
                         constantMultiSet.addCount(ColorSequence { newColorSequence, expr->getColorMaxes() }, expr->getSignedCount());
