+++ src/PetriEngine/ExplicitColored/CMakeLists.txt
@@ -3,11 +3,11 @@ set(CMAKE_INCLUDE_CURRENT_DIR ON)
 add_library(ExplicitColored
     ArcCompiler.cpp
     ColoredPetriNetBuilder.cpp
-    ColoredSuccessorGenerator.cpp
+    SuccessorGenerator/ColoredSuccessorGenerator.cpp
     Algorithms/ExplicitWorklist.cpp
     ColoredResultPrinter.cpp
-    GuardCompiler.cpp
-    GammaQueryCompiler.cpp
+    ExpressionCompilers/GuardCompiler.cpp
+    ExpressionCompilers/GammaQueryCompiler.cpp
     ColoredPetriNet.cpp
 )
